+++ waterfall-cad/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to the
 ## Unreleased
 
 - Fixed an issue in `Waterfall.Transforms` where negative coefficients to `scale` would produce a broken `Solid`
+- Added optics derived from the `Transformable` typeclass, `_translated`, `_scaled`, `_uScaled`, `_rotated`, `_mirrored` and their 2D equivalents.
 
 ## 0.6.2.1